Updated marker modal
Updated marker modal that appears over the map on the landing page when an icon is clicked. Added in that the user must be logged in to see the join button on the marker modal. added in the axios request to add a clean_up_id to a user in the database. Also added in pop-up windows for when the user clicks on the join button so they will see if they have been added to the clean up or if they have already joined the clean up.